app/requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
Werkzeug
markdown
pyotp
WTForms
bleach
qrcode
pillow
Flask-bcrypt
PyCryptodome
gunicorn